拆分物联网
